[REFACTOR] JWT property 값 주입
