+++ src/main/java/com/example/whiplash/config/security/jwt/JwtProperties.java
@@ -2,25 +2,20 @@ package com.example.whiplash.config.security.jwt;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 @Getter
 @Setter
 @Component
-@ConfigurationProperties(prefix = "jwt")
 public class JwtProperties {
 
+    @Value("${jwt.secret}")
     private String secret;
+    @Value("${jwt.access-token-expiration}")
     private long accessTokenExpiration;
+    @Value("${jwt.refresh-token-expiration}")
     private long refreshTokenExpiration;
-    private String isser;
-
-    // 테스트용 기본 설정
-    public JwtProperties() {
-        this.secret = "404E635266556A586E3272357538782F413F4428472B4B6250645367566B5970";
-        this.accessTokenExpiration = 1800000L; // 30분
-        this.refreshTokenExpiration = 604800000L; // 7일
-        this.isser = "whiplash";
-    }
+    @Value("${jwt.issuer}")
+    private String issuer;
 }
